requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1

# OCR
pytesseract==0.3.10
easyocr==1.7.1
Pillow==10.4.0
opencv-python-headless==4.9.0.80

# ML core
torch==2.2.2
torchvision==0.17.2
numpy==1.26.4

# Helpers
python-multipart==0.0.9
pandas==2.2.2
